backend-chatbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere
qdrant-client==1.7.3
sqlalchemy==2.0.23
psycopg[binary]
python-dotenv==1.0.0
pydantic>=2.9
pydantic-settings==2.1.0
pypdf==3.17.4
langchain-text-splitters==0.0.1
pdfplumber==0.10.4
python-multipart==0.0.6
pytest==7.4.3
requests==2.31.0